backend/pyproject.toml
[project]
name = "deep-research-ppt"
version = "0.1.0"
description = "Deep Research based PPT Auto-Generator"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    # deepagents
    "deepagents>=0.1.0",
    # LangChain
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.1.4",
    "langchain-huggingface>=0.1.0",
    "sentence-transformers>=3.0.0",
    # Document parsing
    "pymupdf>=1.24.0",
    "python-pptx>=1.0.0",
    # Web crawling
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.13.0",
    # Vector DB
    "chromadb>=0.5.0",
    # Utilities
    "aiofiles>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D1", "S101"]
